fix: Refactor column components into separate files
