+++ src/components/IngredientReadView.tsx
@@ -6,194 +6,15 @@ import {
   Title, 
   TextInput, 
   Pill,
-  Menu,
-  ActionIcon,
   Flex,
   Paper,
   Table,
-  Anchor,
-  Badge,
-  Group,
-  Stack,
   Checkbox
 } from '@mantine/core';
-import { IconSearch, IconFilter, IconDots, IconCirclePlus } from '@tabler/icons-react';
+import { IconSearch, IconFilter, IconCirclePlus } from '@tabler/icons-react';
 import { colors, spacing, typography } from '@/designTokens';
 import EmptyState from './common/EmptyState';
-
-// Column type definitions
-export interface BaseColumnConfig {
-  key: string;
-  header: string;
-  type: string; // TODO: Define these as an enum
-}
-
-export interface TextColumnConfig extends BaseColumnConfig {
-  type: 'text';
-  size?: 'xs' | 'sm' | 'md' | 'lg';
-  weight?: 'normal' | 'medium' | 'semibold' | 'bold';
-  color?: string;
-}
-
-export interface LinkColumnConfig extends BaseColumnConfig {
-  type: 'link';
-  size?: 'xs' | 'sm' | 'md' | 'lg';
-  color?: string;
-  urlPrefix?: string;
-}
-
-export interface BulletsColumnConfig extends BaseColumnConfig {
-  type: 'bullets';
-  items: Array<{
-    textKey: string;
-    badgeKey?: string;
-    badgeColor?: string;
-  }>;
-}
-
-export interface MenuColumnConfig extends BaseColumnConfig {
-  type: 'menu';
-  actions: Array<{
-    label: string;
-    action: string;
-    color?: string;
-  }>;
-  onAction: (action: string, recordId: string) => void;
-}
-
-export type ColumnConfig = 
-  | TextColumnConfig 
-  | LinkColumnConfig 
-  | BulletsColumnConfig 
-  | MenuColumnConfig;
-
-// Data value types
-export interface TextValue {
-  text: string;
-}
-
-export interface LinkValue {
-  text: string;
-  url?: string;
-}
-
-export interface BulletValue {
-  text: string;
-  badge?: string | number;
-}
-
-export interface BulletsValue {
-  items: BulletValue[];
-}
-
-export type ColumnValue = TextValue | LinkValue | BulletsValue | null;
-
-// Record interface
-export interface IngredientRecord {
-  id: string;
-  [key: string]: ColumnValue | string;
-}
-
-// Column display components
-function TextColumn({ config, value }: { config: TextColumnConfig; value: TextValue }) {
-  return (
-    <Text 
-      size={config.size || 'sm'} 
-      fw={typography.fontWeight[config.weight || 'normal']}
-      c={config.color || colors.text.primary}
-    >
-      {value.text}
-    </Text>
-  );
-}
-
-function LinkColumn({ config, value }: { config: LinkColumnConfig; value: LinkValue }) {
-  return (
-    <Anchor 
-      size={config.size || 'sm'} 
-      c={config.color || colors.blue[600]}
-      href={value.url || `${config.urlPrefix || '#'}${value.text}`}
-      td="none"
-    >
-      {value.text}
-    </Anchor>
-  );
-}
-
-function BulletsColumn({ config, value }: { config: BulletsColumnConfig; value: BulletsValue }) {
-  return (
-    <Stack gap={spacing.xs}>
-      {value.items.map((item, idx) => (
-        <Group key={idx} gap={spacing.xs}>
-          <Text size="xs" c={colors.text.secondary}>•</Text>
-          <Text size="xs" c={colors.text.secondary}>{item.text}</Text>
-          {item.badge && (
-            <Badge 
-              size="xs" 
-              variant="light" 
-              color="gray"
-              radius={spacing.radius.sm}
-            >
-              {typeof item.badge === 'number' ? `+${item.badge}` : item.badge}
-            </Badge>
-          )}
-        </Group>
-      ))}
-    </Stack>
-  );
-}
-
-function MenuColumn({ config, record }: { config: MenuColumnConfig; record: IngredientRecord }) {
-  return (
-    <Menu shadow="md" width={200}>
-      <Menu.Target>
-        <ActionIcon variant="subtle" color="gray">
-          <IconDots size={16} />
-        </ActionIcon>
-      </Menu.Target>
-      <Menu.Dropdown>
-        {config.actions.map((action) => (
-          <Menu.Item 
-            key={action.action}
-            color={action.color}
-            onClick={() => config.onAction(action.action, record.id)}
-          >
-            {action.label}
-          </Menu.Item>
-        ))}
-      </Menu.Dropdown>
-    </Menu>
-  );
-}
-
-// Column renderer component
-function ColumnRenderer({ config, record }: { 
-  config: ColumnConfig; 
-  record: IngredientRecord;
-}) {
-  const value = record[config.key] as ColumnValue;
-
-  if (!value && config.type !== 'menu') {
-    return <Text size="sm" c={colors.text.secondary}>—</Text>;
-  }
-
-  switch (config.type) {
-    case 'text':
-      return <TextColumn config={config as TextColumnConfig} value={value as TextValue} />;
-    
-    case 'link':
-      return <LinkColumn config={config as LinkColumnConfig} value={value as LinkValue} />;
-    
-    case 'bullets':
-      return <BulletsColumn config={config as BulletsColumnConfig} value={value as BulletsValue} />;
-    
-    case 'menu':
-      return <MenuColumn config={config as MenuColumnConfig} record={record} />;
-    
-    default:
-      return <Text size="sm">{String(value)}</Text>;
-  }
-}
+import { ColumnConfig, IngredientRecord, ColumnRenderer } from './columns';
 
 // Main component props
 interface IngredientReadViewProps {
