+++ src/components/columns/TextColumn.tsx
@@ -1,0 +1,20 @@
+import { Text } from '@mantine/core';
+import { colors, typography } from '@/designTokens';
+import { TextColumnConfig, TextValue } from './types';
+
+interface TextColumnProps {
+  config: TextColumnConfig;
+  value: TextValue;
+}
+
+export function TextColumn({ config, value }: TextColumnProps) {
+  return (
+    <Text 
+      size={config.size || 'sm'} 
+      fw={typography.fontWeight[config.weight || 'normal']}
+      c={config.color || colors.text.primary}
+    >
+      {value.text}
+    </Text>
+  );
+}
