+++ src/components/columns/index.ts
@@ -1,0 +1,9 @@
+// Export all column types and interfaces
+export * from './types';
+
+// Export column components
+export { TextColumn } from './TextColumn';
+export { LinkColumn } from './LinkColumn';
+export { BulletsColumn } from './BulletsColumn';
+export { MenuColumn } from './MenuColumn';
+export { ColumnRenderer } from './ColumnRenderer';
